fix : id를 Long에서 Integer로

+++ src/main/java/com/example/kbolog/Request/WatchingRequest.java
@@ -7,12 +7,12 @@ import lombok.Setter;
 @Setter
 public class WatchingRequest {
 
-    private Long gameId;
+    private Integer gameId;
     private String title;
     private String content;
     private String location;
 
-    public WatchingRequest(Long gameId, String title, String content, String location) {
+    public WatchingRequest(Integer gameId, String title, String content, String location) {
         this.gameId = gameId;
         this.title = title;
         this.content = content;
