+++ src/main/java/com/example/kbolog/repository/HitterRepository.java
@@ -9,11 +9,11 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface HitterRepository extends JpaRepository<Hitter, Long> {
+public interface HitterRepository extends JpaRepository<Hitter, Integer> {
 
     //    players - 타자 목록
     @Query(value = "SELECT * FROM hitter h WHERE h.player_id = :playerId", nativeQuery = true)
-    Optional<Hitter> findByPlayer(@Param("playerId") Long playerId);
+    Optional<Hitter> findByPlayer(@Param("playerId") Integer playerId);
 
 }
