+++ src/main/java/com/example/kbolog/repository/PlayerRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface PlayerRepository extends JpaRepository<Player, Long> {
+public interface PlayerRepository extends JpaRepository<Player, Integer> {
     //    아이디로 Player 찾기
     @Query(value = "SELECT * FROM player p WHERE p.player_id = :playerId", nativeQuery = true)
-    Player findByPlayerId(@Param("playerId") Long playerId);
+    Player findByPlayerId(@Param("playerId") Integer playerId);
 
     //    players - 플레이어 리스트
     @Query(value = "SELECT * FROM player", nativeQuery = true)
