+++ src/main/java/com/example/kbolog/repository/WatchingRepository.java
@@ -10,28 +10,28 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.Optional;
 
-public interface WatchingRepository extends JpaRepository<Watching, Long> {
+public interface WatchingRepository extends JpaRepository<Watching, Integer> {
 
 // dashboard - 사용자 관람 기록(로그) ✅
     @Query(value = "SELECT * FROM watching WHERE user_id = :userId", nativeQuery = true)
-    List<Watching> findByUser(@Param("userId") Long userId);
+    List<Watching> findByUser(@Param("userId") Integer userId);
 
 // log - 중복 로그 있나 확인 ✅
     @Query(value = "SELECT EXISTS (SELECT 1 FROM watching WHERE game_id = :gameId AND user_id = :userId)",
             nativeQuery = true)
-    Integer existsByGameAndUser(@Param("gameId") Long gameId, @Param("userId") Long userId);
+    Integer existsByGameAndUser(@Param("gameId") Integer gameId, @Param("userId") Integer userId);
 
 
 
     @Query(value = "SELECT * FROM watching WHERE watching_id = :id", nativeQuery = true)
-    Optional<Watching> findByWatchingId(@Param("id") Long id);
+    Optional<Watching> findByWatchingId(@Param("id") Integer id);
 
 //    log - 로그 작성 ✅
     @Transactional
     @Modifying
     @Query(value = "INSERT INTO watching (user_id, game_id, title, content, location) VALUES (:userId, :gameId, :title, :content, :location)", nativeQuery = true)
-    void write(@Param("userId") Long userId,
-              @Param("gameId") Long gameId,
+    void write(@Param("userId") Integer userId,
+              @Param("gameId") Integer gameId,
               @Param("title") String title,
               @Param("content") String content,
               @Param("location") String location
@@ -41,8 +41,8 @@ public interface WatchingRepository extends JpaRepository<Watching, Long> {
     @Transactional
     @Modifying
     @Query(value = "UPDATE watching SET game_id = :gameId, title = :title, content = :content, location = :location WHERE watching_id = :id", nativeQuery = true)
-    void edit(@Param("id") Long id,
-              @Param("gameId") Long gameId,
+    void edit(@Param("id") Integer id,
+              @Param("gameId") Integer gameId,
               @Param("title") String title,
               @Param("content") String content,
               @Param("location") String location
@@ -52,5 +52,5 @@ public interface WatchingRepository extends JpaRepository<Watching, Long> {
     @Transactional
     @Modifying
     @Query(value = "DELETE FROM watching WHERE watching_id = :id", nativeQuery = true)
-    void delete(@Param("id") Long id);
+    void delete(@Param("id") Integer id);
 }
